requirements.txt
kaggle
notebook
pandas
geopy
faker
mysql-connector-python
seaborn
elasticsearch
Flask 